backend/requirements.txt
# =============================================================================
# FastAPI & Web Framework
# =============================================================================
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# =============================================================================
# Database
# =============================================================================
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1
greenlet==3.0.3

# =============================================================================
# Redis
# =============================================================================
redis==5.0.1
aioredis==2.0.1

# =============================================================================
# Celery (Task Queue)
# =============================================================================
celery[redis]==5.3.6
flower==2.0.1

# =============================================================================
# MinIO / S3
# =============================================================================
minio==7.2.3
boto3==1.34.14

# =============================================================================
# AMI (Asterisk Manager Interface)
# =============================================================================
panoramisk==1.4

# =============================================================================
# File Processing
# =============================================================================
pandas==2.1.4
openpyxl==3.1.2
python-magic==0.4.27
pydub==0.25.1

# =============================================================================
# Phone Number Validation
# =============================================================================
phonenumbers==8.13.27

# =============================================================================
# WebSocket
# =============================================================================
websockets==12.0

# =============================================================================
# Utilities
# =============================================================================
httpx==0.26.0
tenacity==8.2.3
structlog==24.1.0
python-dotenv==1.0.0
cryptography==41.0.7

# =============================================================================
# Testing
# =============================================================================
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0
factory-boy==3.3.0
